+++ src/test/java/com/ixayda/iam/ApplicationIntegrationTest.java
@@ -1,0 +1,13 @@
+package com.ixayda.iam;
+
+import org.springframework.boot.micrometer.metrics.test.autoconfigure.AutoConfigureMetrics;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.context.annotation.Import;
+
+@AutoConfigureMetrics
+@Import(TestcontainersConfiguration.class)
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+abstract class ApplicationIntegrationTest {
+
+}