+++ src/test/java/com/ixayda/iam/IamCoreServiceApplicationTests.java
@@ -8,16 +8,9 @@ import java.net.http.HttpResponse;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.micrometer.metrics.test.autoconfigure.AutoConfigureMetrics;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.context.annotation.Import;
 
-@AutoConfigureMetrics
-@Import(TestcontainersConfiguration.class)
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-class IamCoreServiceApplicationTests {
+class IamCoreServiceApplicationTests extends ApplicationIntegrationTest {
 
 	@LocalServerPort
 	private int port;